requirements.txt
python-dotenv
pypdf
pandas
pyarrow
tiktoken
langchain
langchain-community
langchain-openai
langchain-text-splitters
streamlit
chromadb
